chore: improve constants usage
