--- src/constants.ts
@@ -1,11 +1,0 @@
-export const DEFAULT_USER_IMAGE_URL =
-  'https://cdn.stats.fm/file/statsfm/images/placeholders/users/private.webp';
-
-export const PAGES_WITH_CUSTOM_OGP = [
-  '/user/[id]/[[...deeplink]]',
-  '/artist/[id]',
-  '/track/[id]',
-  '/album/[id]',
-];
-
-export const PAGES_WITH_BASE_LAYOUT = ['/invite/[referrerId]'];
