+++ src/constants/analytics.ts
@@ -1,0 +1,7 @@
+export const ANALYTICS_EVENTS = {
+  // Referral Events
+  REFERRAL: {
+    CTA_CLICKED: 'referral_page_cta_clicked',
+    PAGE_VIEWD: 'referral_child_referral_page_viewed',
+  },
+} as const;
