+++ src/constants/index.ts
@@ -1,0 +1,21 @@
+export const DEFAULT_USER_IMAGE_URL =
+  'https://cdn.stats.fm/file/statsfm/images/placeholders/users/private.webp';
+
+export const PAGES_WITH_CUSTOM_OGP = [
+  '/user/[id]/[[...deeplink]]',
+  '/artist/[id]',
+  '/track/[id]',
+  '/album/[id]',
+];
+
+export const PAGES_WITH_BASE_LAYOUT = ['/invite/[referrerId]'];
+
+export const STORE_APPLE_URL =
+  'https://apps.apple.com/app/spotistats-for-spotify/id1526912392';
+export const STORE_GOOGLE_URL =
+  'https://play.google.com/store/apps/details?id=dev.netlob.spotistats';
+
+export const STORE_APPLE_LOGO =
+  'https://cdn.stats.fm/file/statsfm/images/brands/stores/app_store.webp';
+export const STORE_GOOGLE_LOGO =
+  'https://cdn.stats.fm/file/statsfm/images/brands/stores/play_store.webp';
